feat(engine): cross-shard shared limits + sharded conv aggregation (P2b/P2c)
SharedLimits (Arc atomics): a sharded rule's max_throttle / max_instances /
max_memory_bytes are enforced collectively across all shards — shared sliding
window throttle, exact CAS instance reservation, rule-wide FailRule latch;
rule_instances gauge is now a delta sum across shards. with_limits_shared builds
a shard sharing one SharedLimits; shards=1 keeps the per-machine path.

Sharded conv rules: fixed-window conv rules compile RulePlan.conv_window
(ConvWindowPlan: bucket over + scope keys) and become shardable. Raw qualifying
closes are aggregated across shards in a new ConvStageTask (transform operator):
per-shard watermark barrier seals a bucket only after every shard passes its
end, apply_conv runs over the merged batch (global top-N/sort), shared throttle
applies, then emit. EOS/drained flush is the correct exit; cancel drops unsealed
(partial) buckets; a 30s stale barrier drops stuck buckets with a warning.

Fixes: conv stage now honors on_exceed (FailRule latches, not silent Throttle);
max_instances exact under sharding (DropOldest evicts local + re-reserves,
rejects when budget held by other shards); conv-sink sends one batch per
processed batch instead of per event.

Chore: workspace passes cargo clippy --all-targets --all-features -- -D warnings
(cleared pre-existing toolchain lints; intentional boxed instance-state
collections allowed); version 1.0.0; changelog + tests for the above.

+++ crates/wf-engine/src/match_engine/executor/alert.rs
@@ -29,9 +29,9 @@ pub(crate) fn format_nanos_utc(nanos: i64) -> String {
     let mut s = String::with_capacity(24);
     push_digits(&mut s, year as u32, 4);
     s.push('-');
-    push_digits(&mut s, month as u32, 2);
+    push_digits(&mut s, month, 2);
     s.push('-');
-    push_digits(&mut s, day as u32, 2);
+    push_digits(&mut s, day, 2);
     s.push('T');
     push_digits(&mut s, hour as u32, 2);
     s.push(':');
@@ -321,8 +321,8 @@ mod format_tests {
             1_000_000,
             999_999_999,
             1_000_000_000,
-            86_399_999_999_999_999, // last ms of 1970-01-01
-            86_400_000_000_000_000, // 1970-01-02T00:00:00
+            86_399_999_999_999_999,    // last ms of 1970-01-01
+            86_400_000_000_000_000,    // 1970-01-02T00:00:00
             1_583_020_800_000_000_000, // 2020-03-01 (leap year boundary)
             1_609_459_200_000_000_000, // 2021-01-01
             4_102_444_800_000_000_000, // 2100-01-01 (non-leap century)
